Specify the base in js/parseInt to avoid leading zeroes as octal
Apparently the behavior of this function changed, it used to interpret
leading zeroes as octal, but now assumes base 10 and ignores leading
zeroes. The second argument forces a specific base, to ensure the same
behavior in older browsers (and rhino).

That means I can delete that (probably) slow, ugly abomination of a
helper function. Woohoo. Should've googled better in the first place,
though.