Handle reduced, require init in walk-reduce fns
Made reduce*, which returns a reduced value still wrapped in reduced,
then check for reduced after each step in the walk-reduce fns.
Add tests that failed before this change.
No changes needed for walk-transduce fns.

Change the reduce-fns to require the caller to supply an init value
as using (rf) is not very useful and can't be changed later.