Add new modifiers to for, un-deprecate for variations
Add the modifiers :do, :when-let and :while-let to all for
variations.

Un-deprecate the variations and remove mentions of the :into
stuff. after some hammock time I now think it doesn't really make
sense. It won't be removed, but should be considered deprecated.

Hold off on releasing a new version with this until both tests and a blog post
are written.